--- src/main/java/com/example/infinite/domain/payment/enums/SubscriptionStatus.java
@@ -1,7 +1,0 @@
-package com.example.infinite.domain.payment.enums;
-
-public enum SubscriptionStatus {
-    ACTIVE,       // 구독 활성
-    EXPIRED,      // 기간 만료
-    CANCELLED     // 유저 해지
-}
