--- src/main/java/com/example/infinite/domain/subscriptionmembership/dto/SubscriptionMembershipDto.java
@@ -1,4 +1,0 @@
-package com.example.infinite.domain.subscriptionmembership.dto;
-
-public class SubscriptionMembershipDto {
-}
