--- src/main/java/com/example/infinite/domain/subscriptionmembership/entity/SubscriptionMembership.java
@@ -1,23 +1,0 @@
-package com.example.infinite.domain.subscriptionmembership.entity;
-
-import com.example.infinite.global.common.entity.BaseEntity;
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.hibernate.annotations.SQLDelete;
-import org.hibernate.annotations.SQLRestriction;
-
-@Getter
-@Entity
-@Table(name = "subscription_memberships")
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@SQLDelete(sql = "UPDATE subscription_memberships SET deleted_at = current_timestamp WHERE id = ?")
-@SQLRestriction("deleted_at IS NULL")
-public class SubscriptionMembership extends BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-}
