+++ src/main/java/com/example/infinite/domain/subscriptionmembership/enums/SubscriptionStatus.java
@@ -1,0 +1,7 @@
+package com.example.infinite.domain.subscriptionmembership.enums;
+
+public enum SubscriptionStatus {
+    ACTIVE,
+    EXPIRED,
+    CANCELLED
+}
