+++ src/main/java/com/example/infinite/domain/subscriptionmembership/repository/FanMembershipRepository.java
@@ -1,7 +1,7 @@
-package com.example.infinite.domain.payment.repository;
+package com.example.infinite.domain.subscriptionmembership.repository;
 
-import com.example.infinite.domain.payment.entity.FanMembership;
-import com.example.infinite.domain.payment.enums.SubscriptionStatus;
+import com.example.infinite.domain.subscriptionmembership.entity.FanMembership;
+import com.example.infinite.domain.subscriptionmembership.enums.SubscriptionStatus;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -10,9 +10,7 @@ import java.util.Optional;
 
 public interface FanMembershipRepository extends JpaRepository<FanMembership, Long> {
 
-    // 특정 아티스트에 대한 유저의 활성 멤버십 조회
     Optional<FanMembership> findByUserIdAndArtistIdAndStatus(Long userId, Long artistId, SubscriptionStatus status);
 
-    // 유저의 멤버십 전체 이력 페이징 조회
     Page<FanMembership> findByUserId(Long userId, Pageable pageable);
 }
