--- src/main/java/com/example/infinite/domain/subscriptionmembership/repository/SubscriptionMembershipRepository.java
@@ -1,4 +1,0 @@
-package com.example.infinite.domain.subscriptionmembership.repository;
-
-public class SubscriptionMembershipRepository {
-}
